requirements.txt
stripe
django-environ
django-widget-tweaks